+++ cmd/darkman/main.go
@@ -52,19 +52,31 @@ var toggleCmd = &cobra.Command{
 	},
 }
 
-var runCmd = &cobra.Command{
-	Use:   "run",
-	Short: "Run the darkman service",
-	Long: `This command starts the darkman service itself. It should only
+func newRunCmd() *cobra.Command {
+	var readyFdRaw uint
+	cmd := &cobra.Command{
+		Use:   "run",
+		Short: "Run the darkman service",
+		Long: `This command starts the darkman service itself. It should only
 be used by a service manager, by a  session init script or alike.
 
 The service will run in foreground.`,
-	RunE: func(cmd *cobra.Command, args []string) error {
-		// Avoid showing usage if service fails to start.
-		// See: https://github.com/spf13/cobra/issues/340#issuecomment-374617413
-		cmd.SilenceUsage = true
-		return darkman.ExecuteService()
-	},
+		RunE: func(cmd *cobra.Command, args []string) error {
+			var readyFd *os.File
+			if cmd.Flags().Changed("ready-fd") {
+				readyFd = os.NewFile(uintptr(readyFdRaw), "ready-fd")
+			} else {
+				readyFd = nil
+			}
+
+			// Avoid showing usage if service fails to start.
+			// See: https://github.com/spf13/cobra/issues/340#issuecomment-374617413
+			cmd.SilenceUsage = true
+			return darkman.ExecuteService(readyFd)
+		},
+	}
+	cmd.Flags().UintVar(&readyFdRaw, "ready-fd", 0, "File descriptor for readiness notification")
+	return cmd
 }
 
 var checkCmd = &cobra.Command{
@@ -80,7 +92,7 @@ func init() {
 	rootCmd.AddCommand(setCmd)
 	rootCmd.AddCommand(getCmd)
 	rootCmd.AddCommand(toggleCmd)
-	rootCmd.AddCommand(runCmd)
+	rootCmd.AddCommand(newRunCmd())
 	rootCmd.AddCommand(checkCmd)
 }